Add CHANGELOG.md to WPKit theme
A CHANGELOG.md file has been added to the WPKit theme. This file documents notable changes in the project. Changes include an update to the WPKit theme version, defining a 'WPKIT_VERSION' constant for standardizing the versioning system, modifications to the loop processing posts, and code cleanup in the header.php file.